Handle invalid VSI signal strength values
When some stations as misconfigured and do not send the signal strength
correctly, this results in a `NumberFormatException` which clogs the
logs and makes it difficult to read them. Due to the raised exception,
the message is also ignored by the system.

By defaulting to a ridiculously low value, the message is accounted for
by the system, while making it clear that something is wrong with the
station.