Introduce Gradle build
This allows building the project with Gradle instead of maven. This is
particularly interesting because the AisLib project is not pushed to a
proper maven repository, which makes it impractical to use the HEAD in
other projects.

By having a Gradle build, it is easy to use the AisLib git repository as
a submodule and depend on it in a Gradle composite build.